:bug: Fix notion regex  & UI miscs
